+++ NEWS.rst
@@ -6,7 +6,10 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Patch: Fix using checkout in variable reader.
+* Patch: Improve logging output.
+* Minor: Added verbose flag to produce more logging on stdout
+
 
 3.0.0
 -----